requirements.txt
Pillow==9.5.0
Pillow==10.2.0
